# docs/mkdocs/openapi.yaml
openapi: 3.1.0
info:
  title: Graph Builder API
  description:
    An API to incrementally build and export an RDF graph from various
    sources.
  version: 1.0.0
paths:
  /:
    get:
      summary: Check API status and graph size
      description: Returns the current status and the number of triples in the graph.
      operationId: read_root__get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
  /graph/add:
    post:
      summary: Add Turtle data to the graph
      description: Parses and adds new triples to the in-memory graph.
      operationId: add_to_graph_graph_add_post
      requestBody:
        content:
          text/plain:
            schema:
              type: string
              title: Rdf Data
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /graph/import_tsv:
    post:
      summary: Import data from a TSV file
      description: "Accepts a tsv file upload (TSV), transforms it to RDF

        using a JSON mapping, and adds it to the graph."
      operationId: import_tsv_graph_import_tsv_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Body_import_tsv_graph_import_tsv_post"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /graph/export:
    get:
      summary: Export the entire graph
      description: Serializes and returns the entire graph in the specified format.
      operationId: export_graph_graph_export_get
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            default: turtle
            title: Format
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /graph/clear:
    delete:
      summary: Clear all triples from the graph
      description: Resets the graph to be empty.
      operationId: clear_graph_graph_clear_delete
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    Body_import_tsv_graph_import_tsv_post:
      properties:
        file:
          type: string
          format: binary
          title: File
        mapping_json:
          type: string
          title: Mapping Json
      type: object
      required:
        - file
        - mapping_json
      title: Body_import_tsv_graph_import_tsv_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError